src/pythonDemo/requirements.txt
ontology-python-sdk==0.1.17